@@ -1,55 +1,54 @@
 import Foundation
 
-internal enum MessageKind
-{
-	case Request
-	case Response
-	case Event
+internal enum MessageKind {
+	case request
+	case response
+	case event
 
-	public static func FromString(_ string: String) throws -> MessageKind {
+	public static func fromString(_ string: String) throws -> MessageKind {
 		switch string {
 		case "event":
-			return .Event
+			return .event
 		case "request":
-			return .Request
+			return .request
 		case "response":
-			return .Response
+			return .response
 		default:
-			throw WKInteropError.InvalidMessageKindString
+			throw WKInteropError.invalidMessageKindString
 		}
 	}
 
 	public func toString() -> String {
 		switch self {
-		case .Event:
+		case .event:
 			return "event"
-		case .Request:
+		case .request:
 			return "request"
-		case .Response:
+		case .response:
 			return "response"
 		}
 	}
 }
 
-internal struct Message
-{
-	public static func From(route: String, kind: MessageKind) -> Message {
+internal struct Message: Sendable {
+	public static func from(route: String, kind: MessageKind) -> Message {
 		return Message(id: NSUUID().uuidString, route: route, kind: kind, content: nil)
 	}
 
-	public static func From(route: String, kind: MessageKind, content: WebKitJObject?) -> Message {
+	public static func from(route: String, kind: MessageKind, content: WebKitJObject?) -> Message {
 		return Message(id: NSUUID().uuidString, route: route, kind: kind, content: content)
 	}
 
-	public func toJsonString() -> String
-	{
-		let dictionary = [
-			"id": id,
-			"route": route,
-			"kind": kind.toString(),
-			"content" : UnwrapJObject(content),
-			] as [String : Any]
-		let data = try! JSONSerialization.data(withJSONObject: dictionary, options: JSONSerialization.WritingOptions(rawValue: 0))
+	public func toJsonString() throws -> String {
+		let dictionary =
+			[
+				"id": id,
+				"route": route,
+				"kind": kind.toString(),
+				"content": unwrapJObject(content),
+			] as [String: Any]
+		let data = try JSONSerialization.data(
+			withJSONObject: dictionary, options: JSONSerialization.WritingOptions(rawValue: 0))
 		return String(data: data, encoding: String.Encoding.utf8)!
 	}
 
