@@ -1,18 +1,28 @@
 import Foundation
 import WebKit
 
-internal extension WKScriptMessage
-{
+extension WKScriptMessage {
 	func parse() throws -> Message {
 		let dict = self.body as? NSDictionary
-		if (dict == nil) {
-			throw WKInteropError.InvalidMessage
+		if dict == nil {
+			throw WKInteropError.invalidMessage
 		}
 
-		let id = dict!["id"] as! String
-		let route = dict!["route"] as! String
+		guard let id = dict!["id"] as? String else {
+			throw WKInteropError.invalidMessage
+		}
+
+		guard let route = dict!["route"] as? String else {
+			throw WKInteropError.invalidMessage
+		}
+
+		guard let kindString = dict!["kind"] as? String else {
+			throw WKInteropError.invalidMessage
+		}
+
+		let kind = try MessageKind.fromString(kindString)
+
 		let content = dict!["content"]
-		let kind = try MessageKind.FromString(dict!["kind"] as! String)
-		return Message(id: id, route: route, kind: kind, content: WrapJObject(content))
+		return Message(id: id, route: route, kind: kind, content: wrapJObject(content))
 	}
 }
